--- examples/time_in_zone/data/checkout/checkout.json
@@ -1,112 +1,0 @@
-[
-    [
-        [
-            685,
-            113
-        ],
-        [
-            845,
-            110
-        ],
-        [
-            888,
-            188
-        ],
-        [
-            879,
-            294
-        ],
-        [
-            912,
-            415
-        ],
-        [
-            925,
-            344
-        ],
-        [
-            981,
-            511
-        ],
-        [
-            944,
-            620
-        ],
-        [
-            728,
-            640
-        ]
-    ],
-    [
-        [
-            964,
-            101
-        ],
-        [
-            1124,
-            332
-        ],
-        [
-            1205,
-            480
-        ],
-        [
-            1146,
-            588
-        ],
-        [
-            1280,
-            558
-        ],
-        [
-            1280,
-            97
-        ]
-    ],
-    [
-        [
-            402,
-            137
-        ],
-        [
-            499,
-            138
-        ],
-        [
-            471,
-            209
-        ],
-        [
-            494,
-            312
-        ],
-        [
-            479,
-            419
-        ],
-        [
-            465,
-            378
-        ],
-        [
-            434,
-            540
-        ],
-        [
-            478,
-            648
-        ],
-        [
-            299,
-            641
-        ],
-        [
-            226,
-            539
-        ],
-        [
-            279,
-            387
-        ]
-    ]
-]
