--- examples/time_in_zone/data/traffic/traffic.json
@@ -1,218 +1,0 @@
-[
-    [
-        [
-            1158,
-            326
-        ],
-        [
-            1061,
-            372
-        ],
-        [
-            981,
-            427
-        ],
-        [
-            903,
-            492
-        ],
-        [
-            832,
-            583
-        ],
-        [
-            784,
-            671
-        ],
-        [
-            758,
-            720
-        ],
-        [
-            1097,
-            720
-        ],
-        [
-            1112,
-            620
-        ],
-        [
-            1162,
-            507
-        ],
-        [
-            1231,
-            429
-        ],
-        [
-            1259,
-            404
-        ]
-    ],
-    [
-        [
-            317,
-            203
-        ],
-        [
-            404,
-            201
-        ],
-        [
-            390,
-            260
-        ],
-        [
-            363,
-            322
-        ],
-        [
-            321,
-            403
-        ],
-        [
-            261,
-            501
-        ],
-        [
-            194,
-            621
-        ],
-        [
-            131,
-            720
-        ],
-        [
-            0,
-            720
-        ],
-        [
-            0,
-            609
-        ],
-        [
-            62,
-            537
-        ],
-        [
-            123,
-            467
-        ],
-        [
-            181,
-            397
-        ],
-        [
-            234,
-            327
-        ],
-        [
-            278,
-            264
-        ]
-    ],
-    [
-        [
-            144,
-            706
-        ],
-        [
-            249,
-            530
-        ],
-        [
-            331,
-            391
-        ],
-        [
-            371,
-            306
-        ],
-        [
-            393,
-            253
-        ],
-        [
-            406,
-            202
-        ],
-        [
-            486,
-            199
-        ],
-        [
-            507,
-            260
-        ],
-        [
-            503,
-            369
-        ],
-        [
-            475,
-            504
-        ],
-        [
-            445,
-            620
-        ],
-        [
-            411,
-            720
-        ],
-        [
-            133,
-            720
-        ]
-    ],
-    [
-        [
-            517,
-            201
-        ],
-        [
-            602,
-            192
-        ],
-        [
-            652,
-            261
-        ],
-        [
-            689,
-            366
-        ],
-        [
-            722,
-            511
-        ],
-        [
-            737,
-            635
-        ],
-        [
-            745,
-            720
-        ],
-        [
-            504,
-            720
-        ],
-        [
-            532,
-            568
-        ],
-        [
-            548,
-            443
-        ],
-        [
-            551,
-            358
-        ],
-        [
-            549,
-            272
-        ]
-    ]
-]
