+++ examples/time_in_zone/ultralytics_stream_example.py
@@ -3,8 +3,10 @@ from typing import List
 
 import cv2
 import numpy as np
+from inference import InferencePipeline
+from inference.core.interfaces.camera.entities import VideoFrame
 from ultralytics import YOLO
-from utils.general import find_in_list, get_stream_frames_generator, load_zones_config
+from utils.general import find_in_list, load_zones_config
 from utils.timers import ClockBasedTimer
 
 import supervision as sv
@@ -16,45 +18,34 @@ LABEL_ANNOTATOR = sv.LabelAnnotator(
 )
 
 
-def main(
-    rtsp_url: str,
-    zone_configuration_path: str,
-    weights: str,
-    device: str,
-    confidence: float,
-    iou: float,
-    classes: List[int],
-) -> None:
-    model = YOLO(weights)
-    tracker = sv.ByteTrack(minimum_matching_threshold=0.5)
-    frames_generator = get_stream_frames_generator(rtsp_url=rtsp_url)
-    fps_monitor = sv.FPSMonitor()
+class CustomSink:
+    def __init__(self, zone_configuration_path: str, classes: List[int]):
+        self.classes = classes
+        self.tracker = sv.ByteTrack(minimum_matching_threshold=0.8)
+        self.fps_monitor = sv.FPSMonitor()
+        self.polygons = load_zones_config(file_path=zone_configuration_path)
+        self.timers = [ClockBasedTimer() for _ in self.polygons]
+        self.zones = None
 
-    frame = next(frames_generator)
-    resolution_wh = frame.shape[1], frame.shape[0]
+    def on_prediction(self, detections: sv.Detections, frame: VideoFrame) -> None:
+        if self.zones is None:
+            resolution_wh = frame.image.shape[1], frame.image.shape[0]
+            self.zones = [
+                sv.PolygonZone(
+                    polygon=polygon,
+                    frame_resolution_wh=resolution_wh,
+                    triggering_anchors=(sv.Position.CENTER,),
+                )
+                for polygon in self.polygons
+            ]
 
-    polygons = load_zones_config(file_path=zone_configuration_path)
-    zones = [
-        sv.PolygonZone(
-            polygon=polygon,
-            frame_resolution_wh=resolution_wh,
-            triggering_anchors=(sv.Position.CENTER,),
-        )
-        for polygon in polygons
-    ]
-    timers = [ClockBasedTimer() for _ in zones]
+        self.fps_monitor.tick()
+        fps = self.fps_monitor.fps
 
-    for frame in frames_generator:
-        fps_monitor.tick()
-        fps = fps_monitor.fps
+        detections = detections[find_in_list(detections.class_id, self.classes)]
+        detections = self.tracker.update_with_detections(detections)
 
-        results = model(frame, verbose=False, device=device, conf=confidence)[0]
-        detections = sv.Detections.from_ultralytics(results)
-        detections = detections[find_in_list(detections.class_id, classes)]
-        detections = detections.with_nms(threshold=iou)
-        detections = tracker.update_with_detections(detections)
-
-        annotated_frame = frame.copy()
+        annotated_frame = frame.image.copy()
         annotated_frame = sv.draw_text(
             scene=annotated_frame,
             text=f"{fps:.1f}",
@@ -63,13 +54,13 @@ def main(
             text_color=sv.Color.from_hex("#000000"),
         )
 
-        for idx, zone in enumerate(zones):
+        for idx, zone in enumerate(self.zones):
             annotated_frame = sv.draw_polygon(
                 scene=annotated_frame, polygon=zone.polygon, color=COLORS.by_idx(idx)
             )
 
             detections_in_zone = detections[zone.trigger(detections)]
-            time_in_zone = timers[idx].tick(detections_in_zone)
+            time_in_zone = self.timers[idx].tick(detections_in_zone)
             custom_color_lookup = np.full(detections_in_zone.class_id.shape, idx)
 
             annotated_frame = COLOR_ANNOTATOR.annotate(
@@ -87,11 +78,39 @@ def main(
                 labels=labels,
                 custom_color_lookup=custom_color_lookup,
             )
-
         cv2.imshow("Processed Video", annotated_frame)
-        if cv2.waitKey(1) & 0xFF == ord("q"):
-            break
-    cv2.destroyAllWindows()
+        cv2.waitKey(1)
+
+
+def main(
+    rtsp_url: str,
+    zone_configuration_path: str,
+    weights: str,
+    device: str,
+    confidence: float,
+    iou: float,
+    classes: List[int],
+) -> None:
+    model = YOLO(weights)
+
+    def inference_callback(frame: VideoFrame) -> sv.Detections:
+        results = model(frame.image, verbose=False, conf=confidence, device=device)[0]
+        return sv.Detections.from_ultralytics(results).with_nms(threshold=iou)
+
+    sink = CustomSink(zone_configuration_path=zone_configuration_path, classes=classes)
+
+    pipeline = InferencePipeline.init_with_custom_logic(
+        video_reference=rtsp_url,
+        on_video_frame=inference_callback,
+        on_prediction=sink.on_prediction,
+    )
+
+    pipeline.start()
+
+    try:
+        pipeline.join()
+    except KeyboardInterrupt:
+        pipeline.terminate()
 
 
 if __name__ == "__main__":
